chore(op-challenger): address lifecycle review feedback
