+++ op-acceptance-tests/tests/base/withdrawal/withdrawal_test_helper.go
@@ -45,6 +45,7 @@ func newSystem(t devtest.T, gameType gameTypes.GameType, extra ...presets.Option
 func TestWithdrawal(gt *testing.T, gameType gameTypes.GameType, extra ...presets.Option) {
 	t := devtest.ParallelT(gt)
 	if gameType == gameTypes.SuperPermissionedGameType || gameType == gameTypes.SuperCannonKonaGameType {
+		// TODO(#21861): Enable this when kona-node supports superroot_atTimestamp
 		sysgo.SkipOnKonaNode(t, "super-root proposals require op-node superroot RPC")
 	}
 	sys := newSystem(t, gameType, extra...)
