+++ op-challenger/game/fault/contracts/faultdisputegame131.go
@@ -5,6 +5,7 @@ import (
 	_ "embed"
 
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
+	gameTypes "github.com/ethereum-optimism/optimism/op-challenger/game/types"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching/rpcblock"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
 )
@@ -14,6 +15,15 @@ var faultDisputeGameAbi131 []byte
 
 type FaultDisputeGameContract131 struct {
 	FaultDisputeGameContractLatest
+}
+
+func isLegacyGameClosed(ctx context.Context, game DisputeGameContract) (bool, error) {
+	// Legacy games have no separate close cycle.
+	status, err := game.GetStatus(ctx)
+	if err != nil {
+		return false, err
+	}
+	return status != gameTypes.GameStatusInProgress, nil
 }
 
 func (f *FaultDisputeGameContract131) IsClosed(ctx context.Context) (bool, error) {
